Added support for date attribute
Added support for a (optional) date attribute in a release node.
Thanks go's out to Felix for his input and proof of concept code.

Note that the format for the date is "MM/dd/yyyy", the date will be
shown in local format. If the date does not comply to the format the
date will be shown as is.